Refactor X25519 handshake logic for TransportSession
Consolidated and modernized the X25519 handshake implementation:
- Migrated functionality from TcpSessionX25519Extensions.cs to HandshakeExtensions.cs.
- Replaced event-based message handling with a streamlined RequestAsync approach.
- Improved error handling for malformed packets, invalid proofs, and handshake errors.
- Enforced memory security by securely clearing sensitive data.
- Added detailed XML documentation and modernized code style using C# features.
- Removed legacy TcpSessionX25519Extensions.cs file.
